Rename extension to expected format
