+++ src/UnleashClientBundle.php
@@ -11,6 +11,9 @@ use Unleash\Client\Bundle\DependencyInjection\Compiler\CacheServiceResolverCompi
 use Unleash\Client\Bundle\DependencyInjection\Compiler\HttpServicesResolverCompilerPass;
 use Unleash\Client\Strategy\StrategyHandler;
 
+/**
+ * @internal
+ */
 final class UnleashClientBundle extends Bundle
 {
     public function build(ContainerBuilder $container): void
